mcp-server-app-insight/requirements.txt
mcp
requests
python-dotenv
pydantic
pytest
sseclient-py